+++ src/Console/Application.php
@@ -76,6 +76,7 @@ class Application extends BaseApplication
         if (in_array('opcache', $this->config['extensions'], true)) {
             $commands[] = new CacheToolCommand\OpcacheConfigurationCommand();
             $commands[] = new CacheToolCommand\OpcacheResetCommand();
+            $commands[] = new CacheToolCommand\OpcacheResetFileCacheCommand();
             $commands[] = new CacheToolCommand\OpcacheStatusCommand();
             $commands[] = new CacheToolCommand\OpcacheStatusScriptsCommand();
             $commands[] = new CacheToolCommand\OpcacheInvalidateScriptsCommand();